site/pages/menus: Fix broken advanced example links
Previously the "Menus" page of the documentation site listed several
"Responsive Menus That Hide" advanced examples. These are implemented as
links to static pages. However when attempting to click on these in the
browser, the links would not be clickable.

This commit fixes the issue of them not being clickable, by adding an
appropriate event handler which prevents the global 'click' handler from
triggering. Furthermore, it removes the useless source listing for the
list of examples.